social_media_api/requirements.txt
asgiref==3.9.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.4
django-csp==4.0
django-environ==0.12.0
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.0
drf-yasg==1.21.10
Flask==3.1.0
Flask-Negotiate==0.1.0
gunicorn==23.0.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyodbc==5.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
Werkzeug==3.1.3
whitenoise==6.9.0
